refactor: merge AI/import DTOs into CoursePayloadDto
