+++ apps/backend/src/main/java/com/codestar/backend/controller/AiCourseController.java
@@ -2,7 +2,7 @@ package com.codestar.backend.controller;
 
 import com.codestar.backend.dto.ApiResponseDto;
 import com.codestar.backend.dto.ai.GenerateCourseRequestDto;
-import com.codestar.backend.dto.ai.GenerateCourseResponseDto;
+import com.codestar.backend.dto.course.CoursePayloadDto;
 import com.codestar.backend.exception.ApiException;
 import com.codestar.backend.security.AuthenticatedUser;
 import com.codestar.backend.service.AiCourseService;
@@ -33,13 +33,13 @@ public class AiCourseController {
 
     @PostMapping("/generate")
     @PreAuthorize("hasAnyRole('TEACHER', 'ADMIN', 'SUPER_ADMIN')")
-    public ResponseEntity<ApiResponseDto<GenerateCourseResponseDto>> generate(@Valid @RequestBody GenerateCourseRequestDto request, @AuthenticationPrincipal AuthenticatedUser principal) {
+    public ResponseEntity<ApiResponseDto<CoursePayloadDto>> generate(@Valid @RequestBody GenerateCourseRequestDto request, @AuthenticationPrincipal AuthenticatedUser principal) {
         if (!aiRateLimiter.tryAcquire(principal.getId())) {
             audit.event("ai.course.rate_limited").field("actorId", principal.getId()).warn();
             throw ApiException.tooManyRequests("AI rate limit exceeded, please wait before generating another course");
         }
 
-        GenerateCourseResponseDto draft = aiCourseService.generate(principal.getId(), request);
+        CoursePayloadDto draft = aiCourseService.generate(principal.getId(), request);
         return ResponseEntity.ok(new ApiResponseDto<>(true, "Course draft generated", draft));
     }
 }
